requirements.txt
Django==1.5.1
djangorestframework==2.2.5
wsgiref==0.1.2
